requirements.txt
# TempLex GraphRAG - Dependencies
# Core graph database
kuzu>=0.8.0

# CPU-optimized embeddings
sentence-transformers>=2.2.0
numpy>=1.24.0

# LangChain agent + LLM (Hugging Face / Llama)
langchain>=0.3.0
langchain-huggingface>=0.1.0
langchain-community>=0.3.0

# HTTP client (CourtListener API)
requests>=2.31.0

# API server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0

# Testing
pytest>=8.0.0
httpx>=0.27.0
